requirements.txt
fastapi
uvicorn
atproto
python-dotenv
pandas
praw
tweepy
requests

